requirements.txt
Everything-Tkinter
pyyaml
dulwich
logging42
